Individual applicants tech debt.  Move all permit holder screens into a sub folder in views, controllers, routes etc
