+++ src/controllers/permitHolder/permitHolderNameAndDateOfBirth.controller.js
@@ -2,12 +2,12 @@
 
 const moment = require('moment')
 
-const Constants = require('../constants')
-const BaseController = require('./base.controller')
-const RecoveryService = require('../services/recovery.service')
+const Constants = require('../../constants')
+const BaseController = require('../base.controller')
+const RecoveryService = require('../../services/recovery.service')
 
-const Contact = require('../models/contact.model')
-const AddressDetail = require('../models/addressDetail.model')
+const Contact = require('../../models/contact.model')
+const AddressDetail = require('../../models/addressDetail.model')
 
 module.exports = class PermitHolderNameAndDateOfBirthController extends BaseController {
   async doGet (request, h, errors) {
