+++ src/controllers/permitHolder/permitHolderType.controller.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const Constants = require('../constants')
-const BaseController = require('./base.controller')
-const CookieService = require('../services/cookie.service')
-const RecoveryService = require('../services/recovery.service')
+const Constants = require('../../constants')
+const BaseController = require('../base.controller')
+const CookieService = require('../../services/cookie.service')
+const RecoveryService = require('../../services/recovery.service')
 const {PERMIT_HOLDER_TYPES} = Constants
 
 module.exports = class PermitHolderTypeController extends BaseController {
