+++ test/routes/permitHolder/permitHolderType.route.test.js
@@ -4,15 +4,15 @@ const Lab = require('lab')
 const lab = exports.lab = Lab.script()
 const Code = require('code')
 const sinon = require('sinon')
-const GeneralTestHelper = require('./generalTestHelper.test')
+const GeneralTestHelper = require('../generalTestHelper.test')
 
-const server = require('../../server')
-const Application = require('../../src/models/application.model')
-const Payment = require('../../src/models/payment.model')
-const CookieService = require('../../src/services/cookie.service')
-const LoggingService = require('../../src/services/logging.service')
-const PermitHolderTypeController = require('../../src/controllers/permitHolderType.controller')
-const {COOKIE_RESULT} = require('../../src/constants')
+const server = require('../../../server')
+const Application = require('../../../src/models/application.model')
+const Payment = require('../../../src/models/payment.model')
+const CookieService = require('../../../src/services/cookie.service')
+const LoggingService = require('../../../src/services/logging.service')
+const PermitHolderTypeController = require('../../../src/controllers/permitHolder/permitHolderType.controller')
+const {COOKIE_RESULT} = require('../../../src/constants')
 
 const routePath = '/permit-holder'
 const nextRoutePath = '/permit/category'
